GTX-528 - Updated naming to better match confluence page with test pack info, plus fixed all linting warns
