+++ eslint.config.mjs
@@ -3,7 +3,8 @@ import tseslint from 'typescript-eslint';
 import unusedImports from 'eslint-plugin-unused-imports';
 
 export default [
-    { files: ['**/*.{js,mjs,cjs,ts}'] },
+    { files: ['**/*.{js,ts}'] },
+    { ignores: ['playwright-report/*'] },
     { languageOptions: { globals: globals.node } },
     ...tseslint.configs.strict,
     ...tseslint.configs.stylistic,
@@ -12,7 +13,7 @@ export default [
             'unused-imports': unusedImports,
         },
         rules: {
-            'no-unused-vars': 'warn',
+            '@typescript-eslint/no-unused-vars': 'warn',
             'no-undef': 'warn',
             'no-undefined': 'warn',
             'unused-imports/no-unused-imports': 'warn',
