+++ gameClientToGameWindowTests/helpers/conversion.ts
