+++ gameClientToGameWindowTests/helpers/getValuesFromConsoleMessages.ts
