+++ gameClientToGameWindowTests/helpers/getValuesFromPage.ts
