+++ gameClientToGameWindowTests/helpers/index.ts
@@ -1,0 +1,20 @@
+export { launchGame } from './launchGame';
+export {
+    validateConsoleMessages,
+    deletePreviousConsoleMessages,
+} from './validateConsoleMessages';
+export { login } from './login';
+export { startEventListener } from './startEventListener';
+export { messageExists, getIndexOfExpectedMessage } from './messageExists';
+export { spin } from './makeSpin';
+export { makeDeposit } from './makeDeposit';
+export { readGames } from './readGames';
+export { getValueFromConsoleMessages } from './getValuesFromConsoleMessages';
+export { currencyStringToNumber, numberToTwoDecimalPlaces } from './conversion';
+export {
+    getBalanceGameWindow,
+    getStakeAmountGameWindow,
+    getWinAmountGameWindow,
+    getWinLossAmountGameWindow,
+} from './getValuesFromPage';
+export { recoverFromFreeSpins } from './recoverFromFailure';
