+++ gameClientToGameWindowTests/helpers/launchGame.ts
