+++ gameClientToGameWindowTests/helpers/readGames.ts
