+++ gameClientToGameWindowTests/helpers/startEventListener.ts
