+++ gameClientToGameWindowTests/helpers/validateConsoleMessages.ts
