+++ gameClientToGameWindowTests/soundCheck.spec.ts
@@ -6,14 +6,14 @@ import {
     startEventListener,
     readGames,
     deletePreviousConsoleMessages,
-} from './utilities';
+} from './helpers';
 import { ExpectedMessage } from '../types/expectedMessage';
 
 let page: Page;
-readGames().forEach((game) => {
+readGames().forEach((game: string) => {
     const ICON_MUTED_CLASS = 'muted';
 
-    test.describe(`Testing with text: ${game}`, () => {
+    test.describe(`Testing with game: ${game}`, () => {
         test.beforeEach(async ({ browser }) => {
             page = await browser.newPage();
             await login(page);
@@ -23,7 +23,7 @@ readGames().forEach((game) => {
             await page.close();
         });
 
-        test('Test sound toggle', async () => {
+        test('sound toggle', async () => {
             const consoleMessages: string[] = [];
             startEventListener(page, consoleMessages);
             await launchGame(page, game, consoleMessages);
