+++ gameClientToGameWindowTests/supportedCurrencies.spec.ts
@@ -5,12 +5,12 @@ import {
     login,
     startEventListener,
     readGames,
-} from './utilities';
+} from './helpers';
 import { ExpectedMessage } from '../types/expectedMessage';
 
 let page: Page;
-readGames().forEach((game) => {
-    test.describe(`Testing with text: ${game}`, () => {
+readGames().forEach((game: string) => {
+    test.describe(`Testing with game: ${game}`, () => {
         test.beforeEach(async ({ browser }) => {
             page = await browser.newPage();
             await login(page);
@@ -20,12 +20,12 @@ readGames().forEach((game) => {
             await page.close();
         });
 
-        test('gameLaunch', async () => {
+        test('supported currencies', async () => {
             const consoleMessages: string[] = [];
             startEventListener(page, consoleMessages);
             await launchGame(page, game, consoleMessages);
             await validateConsoleMessages(
-                ExpectedMessage.GAME_LOAD_COMPLETE,
+                ExpectedMessage.CURRENCY_GBP,
                 consoleMessages
             );
         });
