+++ playwright.config.ts
@@ -14,7 +14,7 @@ dotenv.config({
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
-    testDir: './skyVegasTests',
+    testDir: './gameClientToGameWindowTests',
     /* Run tests in files in parallel */
     fullyParallel: false,
     /* Fail the build on CI if you accidentally left test.only in the source code. */
