--- skyVegasTests/utilities/index.ts
@@ -1,26 +1,0 @@
-export { launchGame } from './launchGameHelper';
-export {
-    validateConsoleMessages,
-    deletePreviousConsoleMessages,
-} from './validateConsoleMessagesHelper';
-export { login } from './loginHelper';
-export { startEventListener } from './startEventListenerHelper';
-export {
-    messageExists,
-    getIndexOfExpectedMessage,
-} from './messageExistsHelper';
-export { spin } from './makeSpinHelper';
-export { makeDeposit } from './makeDepositHelper';
-export { readGames } from './readGamesHelper';
-export { getValueFromConsoleMessages } from './getValuesFromConsoleMessagesHelper';
-export {
-    currencyStringToNumber,
-    numberToTwoDecimalPlaces,
-} from './conversionHelper';
-export {
-    getBalanceGameWindow,
-    getStakeAmountGameWindow,
-    getWinAmountGameWindow,
-    getWinLossAmountGameWindow,
-} from './getValuesFromPageHelper';
-export { recoverFromFreeSpins } from './recoverFromFailureHelper';
